requirements.txt
fastapi
uvicorn
yfinance
pandas
pydantic
textblob
python-dotenv
requests
streamlit
altair
